elevtor/requirements.txt
defusedxml==0.4.1
django==1.10.3
django-allauth==0.28.0
django-rest-auth==0.8.2
django-rest-swagger==2.1.0
djangorestframework==3.5.3
itypes==1.1.0
oauthlib==2.0.0
openapi-codec==1.1.7
pillow==3.4.2
python3-openid==3.0.10
requests==2.12.1
requests-oauthlib==0.7.0
setuptools==27.2.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
vs2015_runtime==14.0.25123
wheel==0.29.0
coreapi==2.0.9

